ocnN and wgh0 dev

+++ forcing/ocnN/make_forcing_main.py
@@ -4,12 +4,10 @@ the fields from history files for another run.
 
 Designed to run only as backfill.
 
-NOTE: This does not yet do biogeochemical variables.
-
 Testing:
-run make_forcing_main.py -g wgh0 -gtx cas6_traps2_x2b -ro 0 -r backfill -s continuation -d 2017.07.04 -f ocnN -test True
+run make_forcing_main.py -g wgh1 -gtx cas6_traps2_x2b -ro 0 -r backfill -s continuation -d 2017.07.04 -f ocnN -do_bio True -test True
 
-Performance: 1 minute per day on mac with wgh0 grid (1.5 minutes for start_type = new)
+Performance: 6 minutes per day on mac for wgh1 grid with do_bio = True and start_type = new.
 """
 
 from pathlib import Path
@@ -29,9 +27,9 @@ from time import time
 import numpy as np
 from subprocess import Popen as Po
 from subprocess import PIPE as Pi
+import pickle
 
-from lo_tools import Lfun, zfun, zrfun
-import Ofun_nc_xarray
+from lo_tools import Lfun, zfun, zrfun, Ofun_nc
 
 # this directory is created, along with Info and Data subdirectories, by ffun.intro()
 out_dir = Ldir['LOo'] / 'forcing' / Ldir['gridname'] / ('f' + Ldir['date_string']) / Ldir['frc']
@@ -43,8 +41,6 @@ in_dir = Ldir['roms_out'] / Ldir['gtagex'] / ('f' + Ldir['date_string'])
 this_dt = datetime.strptime(Ldir['date_string'], Lfun.ds_fmt)
 
 # list of history files from the original grid to work from (all Path objects)
-# h_list = sorted(in_dir.glob('ocean_his_*'))
-# NEW
 h_list = Lfun.get_fn_list('hourly', Ldir, Ldir['date_string'], Ldir['date_string'])
 
 verbose = False
@@ -69,7 +65,8 @@ for ii in range(NT):
     temp_out_fn_list.append(temp_out_fn)
     this_dir = str(Ldir['LO'] / 'forcing' / Ldir['frc']) + '/'
     cmd_list = ['python', this_dir + 'get_one_time.py', '-grid_fn', grid_fn,
-         '-his_fn', his_fn, '-start_type', Ldir['start_type'], '-out_fn', str(temp_out_fn)]
+         '-his_fn', his_fn, '-start_type', Ldir['start_type'], '-out_fn', str(temp_out_fn),
+         '-do_bio', str(Ldir['do_bio'])]
     proc = Po(cmd_list, stdout=Pi, stderr=Pi)
     proc_list.append(proc)
         
@@ -89,26 +86,77 @@ sys.stdout.flush()
 
 # Write files to NetCDF.
 
+# Write climatology file making use of zrfun.get_varinfo().
 tt0 = time()
 out_fn = out_dir / 'ocean_clm.nc'
 out_fn.unlink(missing_ok=True)
-Ofun_nc_xarray.make_clm_file(temp_out_fn_list, NT, out_fn)
+ds = xr.Dataset()
+# Initialize Dataset and save variable names to loop over
+dd0 = pickle.load(open(temp_out_fn_list[0], 'rb'))
+ddk = dd0.keys()
+for vn in ddk:
+    if vn == 'ocean_time':
+        pass
+    else:
+        vinfo = zrfun.get_varinfo(vn, vartype='climatology')
+        dims = (vinfo['time_name'],) + vinfo['space_dims_tup']
+        ds[vn] = (dims, np.nan * np.ones((NT,) + dd0[vn].shape))
+        ds[vn].attrs['units'] = vinfo['units']
+        ds[vn].attrs['long_name'] = vinfo['long_name']
+# loop over the temp files
+ot_vec = np.nan * np.ones(NT)
+ii = 0
+for temp_out_fn in temp_out_fn_list:
+    dd = pickle.load(open(temp_out_fn, 'rb'))
+    for vn in ddk:
+        ddv = dd[vn]
+        if verbose: # debugging
+            print('\n%d %s max = %0.1f' % (ii,vn,np.nanmax(ddv)))
+        if vn == 'ocean_time':
+            ot_vec[ii] = ddv
+        else:
+            if ddv.ndim == 2:
+                ds[vn][ii, :, :] = ddv
+            elif ddv.ndim == 3:
+                ds[vn][ii, :, :, :] = ddv
+            else:
+                print('problem with ndim?')
+    ii += 1
+# Add the time coordinates
+for vn in ddk:
+    if vn == 'ocean_time':
+        pass
+    else:
+        vinfo = zrfun.get_varinfo(vn, vartype='climatology')
+        tname = vinfo['time_name']
+        # time coordinate
+        ds[tname] = ((tname,), ot_vec)
+        ds[tname].attrs['units'] = Lfun.roms_time_units
+# and save to NetCDF
+Enc_dict = {vn:zrfun.enc_dict for vn in ds.data_vars}
+ds.to_netcdf(out_fn, encoding=Enc_dict)
+ds.close()
 print('- Write clm file: %0.2f sec' % (time()-tt0))
 sys.stdout.flush()
 
-tt0 = time()
-in_fn = out_dir / 'ocean_clm.nc'
-out_fn = out_dir / 'ocean_ini.nc'
-out_fn.unlink(missing_ok=True)
-Ofun_nc_xarray.make_ini_file(in_fn, out_fn)
-print('- Write ini file: %0.2f sec' % (time()-tt0))
-sys.stdout.flush()
+# Write initial condition file if needed
+if Ldir['start_type'] == 'new':
+    tt0 = time()
+    in_fn = out_dir / 'ocean_clm.nc'
+    out_fn = out_dir / 'ocean_ini.nc'
+    out_fn.unlink(missing_ok=True)
+    Ofun_nc.make_ini_file(in_fn, out_fn)
+    print('- Write ini file: %0.2f sec' % (time()-tt0))
+    sys.stdout.flush()
+else:
+    print('- Skipped writing ocean_ini.nc')
 
+# Write boundary file
 tt0 = time()
 in_fn = out_dir / 'ocean_clm.nc'
 out_fn = out_dir / 'ocean_bry.nc'
 out_fn.unlink(missing_ok=True)
-Ofun_nc_xarray.make_bry_file(in_fn, out_fn)
+Ofun_nc.make_bry_file(in_fn, out_fn)
 print('- Write bry file: %0.2f sec' % (time()-tt0))
 sys.stdout.flush()
 
@@ -122,7 +170,10 @@ def print_info(fn):
     ds.close()
 
 # check results
-nc_list = ['ocean_clm.nc', 'ocean_ini.nc', 'ocean_bry.nc']
+if Ldir['start_type'] == 'new':
+    nc_list = ['ocean_clm.nc', 'ocean_ini.nc', 'ocean_bry.nc']
+else:
+    nc_list = ['ocean_clm.nc', 'ocean_bry.nc']
 if verbose:
     # print info about the files to the screen
     for fn in nc_list:
